make cpp friendly

+++ danklib/grid.h
@@ -11,18 +11,18 @@ typedef struct {
     size_t h;
 } grid;
 
-grid grid_init(size_t size, int w, int h);
+grid grid_init(size_t size, size_t w, size_t h);
 
 // buf: size
-void grid_set(grid g, void *buf, int x, int y);
+void grid_set(grid g, void *buf, size_t x, size_t y);
 
 // buf: size
-bool grid_get(grid g, void *buf, int x, int y);
+bool grid_get(grid g, void *buf, size_t x, size_t y);
 
 // buf: size
 // return is neighbour
-bool grid_get_neighbour8(grid g, void *buf, int x, int y, int n);
+bool grid_get_neighbour8(grid g, void *buf, size_t x, size_t y, int n);
 
-bool grid_index_neighbour8(grid g, int x, int y, int i, int *xp, int *yp);
+bool grid_index_neighbour8(grid g, size_t x, size_t y, int i, int *xp, int *yp);
 
 void grid_delete(grid g);
